Keep a reference on the control-point during signal emission
I've run into a problem where my code unrefs a control-point from the
GUPnP::device-proxy-available signal handler. The code looks for a
specific device to appear and the control-point will not be needed any
longer when that device shows up. Unfortunately this leads to a crash in
GUPnPControlPoint. Fortunately this can be fixed easily. A patch is
attached.