Utilize libconic if available
libconic is the library available on maemo to get signaled when a network
connection goes up or down. This patch adds support for that. Since
libconic doesn't tell us which interface went-up and there is only one
connection at a time on maemo platforms, we just create contexts for all
available up interfaces on connection event and destroy all existing
contexts on disconnection event.