Add GList variants of begin/end/send_action
This is mainly done to allow PyGI to be able to call actions.
gobject-introspection do not (yet) support vararg and on top of that
PyGI maintainer refuses to handle reference type 'in' parameters
correctly because the only way to do that is 'inefficient'.

Co-author and reviewer: Zeeshan Ali <zeeshanak@gnome.org>