Port from dbus-glib to GDBus
This drop the dbus-glib dependency but bumps the gio dependency to 2.26 (latest
stable release) if network-manager is to be used as the context manager. One
"issue" here is that GDBus (and gio in general it seems) simply uses the main
context associated with the current thread's context and do not allow us to set
the main context we are told to use by the application. Its not really an issue
since applications can use g_main_context_push_thread_default() instead of
passing the main context to gupnp, however this is still a fundamental change in
the behavior of gupnp.