Ask libsoup to timeout on blocking I/O
Although DLNA requirement 7.2.8.7 specifically states that we need to close the
socket if there is no pending requests on that, the DLNA CTT test case for
7.2.8.7 doesn't reply to our request and then wrongly assumes that we should be
closing the socket after 60 seconds even though there is a pending request on
the socket. Astonishingly the same test case for clients does the right thing.

Until the bug is resolved in DLNA CTT, we can temporarily work around the issue
by setting-up a timeout on blocking I/O as well.