Revert "Ask libsoup to timeout on blocking I/O"
This reverts commit 9da38618ef5ca73183864497a099e51f745ad638.

This is no longer needed as the bug in DLNA CTT has been fixed and
hopefully be available in the next release. The issue turned out to that
CTT *was* replying to all the HTTP requests but in one of them, it didn't
set any headers at all and libsoup was still expecting more data on the
socket and hence didn't close the socket since from its POV the socket was
still active.