RootDevice delays description doc unref to finalize
Each service keeps a pointer to its node in the device description doc so
we could easily end-up service accessing an invalid node pointer in
situations where the service in question is not unrefed BEFORE the dispose
of the associated device is run.

This was the cause of rygel crashing if you hit CTRL^C at its initialization.