Make GUPnPContextManager an abstract base class
gupnp_contxt_manager_new is renamed to gupnp_context_manager_create to
emphasize the factory-method character of it. It now only creates an
instance of the implementation.

The chained listening to the context-unavailable signal is now done by an
object signal handler run in first phase (G_SIGNAL_RUN_FIRST).

Previously, gupnp_context_manager_new was creating an instance of
GUPnPContextManager and the specific implementation, wrapping the
implementation in the GUPnPContextManager instance. The instance was
connecting to the context-unavailable signal of the implementation to take
care of the managed objects.