Signalize higher version services
If a control point was looking for a service instead of a device it was
only responding on exact version matches, but according to UDA also
higher versions are compatible as well.

This patch introduces a compare function that takes the service version
into account instead of a simple string comparison.

https://bugzilla.gnome.org/show_bug.cgi?id=689028