Fix missing event notifications
If a client used the way to respond to a NOTIFY call described in UDA 1.1:

HTTP/1.1 200 OK

and nothing else, libsoup doesn't know that there was no body to come and was
running into a timeout, not sending the other messages queued on this
connection since we request the connection to be kept alive.

By adding "Connection: close" to the NOTIFY message, libsoup will know that
the message is done since the client closes the connection.

If a client misbehaves and doesn't close the connection, we're still having a
problem. This should be considered a work-around and a nicer solution with
libsoup needs to be found.