context: Use ASCII-only user-agent
We were using the application name, which is translated, as the
identifier for our program in the user-agent. This caused some
UPnP servers, such as the one built into Synology NAS to fail
to answer our announcement requests.

Use the program name instead, as this is supposed to be ASCII only.

https://bugzilla.gnome.org/show_bug.cgi?id=704953