Do not try to reuse message on error
If one of the call-backs has an error, just recreate the soup message. If we
want to re-use the old soup message we need to get the reference counting
right etc. This could lead to cases where notifications were not sent properly
because the first URL was failing and the requeuing failed as well.

Also it caused crashes by leaving a dangling pointer in pending_messages.

https://bugzilla.gnome.org/show_bug.cgi?id=703263