service-proxy: unconditionally remove server handler in dispose
The handler installed on the SoupServer owned by the GUPnPContext
needs to be removed when the GUPnPServiceProxy is disposed. This
happens in the unsubscribe method, but this method is only called
unconditionally. If a subscription request is in progress, then
the proxy is not subscribed, but it has already installed a
handler. So make sure that this handler is removed under all
circumstances.

Signed-off-by: Sven Neumann <neumann@teufel.de>

https://bugzilla.gnome.org/show_bug.cgi?id=777491