Avoid a crash in GUPnPServiceProxy::subscription_expire()
Can't explain why, but we are seeing crashes in subscription_expire()
that seem to be caused by an attempt to renew the subscription even
though the SID is not set, leading to a crash in
soup_message_headers_append().

Add an extra check in subscription_expire() and also make sure
that the subscription timeout is removed before a new timeout is
installed in subscription_response().

https://bugzilla.gnome.org/show_bug.cgi?id=780715

Signed-off-by: Sven Neumann <neumann@teufel.de>