service-proxy: Fix endless loop on early notify
According to UDA, the service has to send the initial event AFTER making
sure that the SUBSCRIBE call is returned.

Some devices, such as the media server in Fritz OS 6 or IGD in D-Link
DIR 635 send the initial event state BEFORE sending the answer to
subscibe. If that happens, just ignore the event.

Fixes #8