resource-factory: Version-independent typeregister
Until now the types registered in the GUPnPResourceFactory had to be an
exact match. For example, egistering a type for
urn:schemas-upnp-org:service:AVTransport:2 would only return this type
for AVTransport:2 but not for a detected AVTransport:1. It is now
possible to register for AVTransport only without a concrete version.

https://bugzilla.gnome.org/show_bug.cgi?id=764498