service: Remove constructor
There is no need to do the actions in the constructor, doing it in
constructed is early enough.

Ideally we would even be a GAsyncInitable and wait for the introspection
to finish. That way a whole lot of code in subscription handling etc.
could be just omitted, but this requires chainging API quite intrusively
and shall be done for 2.0 API release.