Device,Service,RootDevice: Fix use after free
Devices and Services created by a RootDevice were keeping a pointer to
their XML element, but not to the document, so it could happen that if
the service lived longer than the RootDevice, the element pointer was
invalid.

This is already implemented like this for DeviceProxy and ServiceProxy.

No idea why the description-doc property was introduced additionally
to the "document" property on DeviceInfo.

Fixes #42