ContextManager: Handle filter events properly
If the filter is changed, announce all changed to contexts as well

This means if the filter is changed, causing a known device not to match
anymore, it will signal "context-unavailable" and drop all managed
devices attached to that context.

Likewise, for filter changes that cause contexts to not being filtered
out any more, "context-available" will be signalled.

Fixes #36
Fixes #37